Implemented more complicated water model.     -   Laid the ground work for edventually generating the water model based on data sent in from constructor.     -   The camera is acting weird tho, need to investigate that.

+++ include/LoadBMP.hpp
@@ -12,5 +12,5 @@
 //  format should be: red, green, blue, alpha
 //  the size of the .bmp file should be equal to the number of pixels times 4
 //  width and height should get their apropiet values(should be found in the header of the .bmp file)
-//  keep in mind that the values of the pixels are stored in byte, which is to say they are not floats
-GLfloat* LoadBMP(std::string path, int& width, int& height);
+//  it turns out that sending the int values is better than converting them to floats
+GLint* LoadBMP(std::string path, int& width, int& height);
